fixing Comments

+++ app/src/main/java/experts/rihanna/appsmatic/com/rihannaexperts/API/ModelsPOJO/Services/ExpertServices/Service.java
@@ -18,13 +18,13 @@ public class Service {
     private Integer serviceId;
     @SerializedName("price")
     @Expose
-    private Integer price;
+    private Double price;
     @SerializedName("discount_amount")
     @Expose
-    private Integer discountAmount;
+    private Double discountAmount;
     @SerializedName("discount_percentage")
     @Expose
-    private Integer discountPercentage;
+    private Double discountPercentage;
     @SerializedName("LastUpdate")
     @Expose
     private String lastUpdate;
@@ -56,27 +56,27 @@ public class Service {
         this.serviceId = serviceId;
     }
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
 
-    public Integer getDiscountAmount() {
+    public Double getDiscountAmount() {
         return discountAmount;
     }
 
-    public void setDiscountAmount(Integer discountAmount) {
+    public void setDiscountAmount(Double discountAmount) {
         this.discountAmount = discountAmount;
     }
 
-    public Integer getDiscountPercentage() {
+    public Double getDiscountPercentage() {
         return discountPercentage;
     }
 
-    public void setDiscountPercentage(Integer discountPercentage) {
+    public void setDiscountPercentage(Double discountPercentage) {
         this.discountPercentage = discountPercentage;
     }
